backend/requirements.txt
fastapi
uvicorn
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlite-vec
openai
numpy
dotenv